Yinsen_M2/requirements.txt
# Core dependencies
pyyaml>=6.0.1
python-dotenv>=1.0.0
#colorlog>=6.7.0
openai>=1.0.0

# backend dependencies
fastapi>=0.115.4
uvicorn>=0.25.0

# Tool usage - ACI.dev
aipolabs

# Audio processing
# Voice input requirements

numpy>=1.24.0
matplotlib>=3.7.2
pandas>=2.0.3
plotly>=5.19.0
kaleido>=0.2.1

# System monitoring
psutil>=5.9.0

# Utility packages
tqdm>=4.66.0
requests>=2.31.0
python-dateutil>=2.8.2
gTTS>=2.5.0
EOL 
